refactor: update MikroORM reflection dependency and improve entity type imports
- Updated @mikro-orm/reflection dependency to version 6.4.16 in package.json and package-lock.json.
- Introduced a new types.ts file to centralize entity type exports, avoiding circular dependencies in entity models.
- Modified user, session, and todo entity imports to reference types from the new types.ts file for better clarity and maintainability.
