+++ src/entities/types.ts
@@ -1,0 +1,4 @@
+// Entity type exports to avoid circular dependencies
+export type User = import('./models/user.entity').User;
+export type Todo = import('./models/todo.entity').Todo;
+export type Session = import('./models/session.entity').Session;
